# Conflicts:
#	src/main/java/cz/jcu/prf/uai/javamugs/App.java

+++ src/main/java/cz/jcu/prf/uai/javamugs/logic/GameReport.java
@@ -32,7 +32,7 @@ public class GameReport {
 	/**
 	 * @return current score multiplier
 	 */
-	public long getMultiplier() {
+	public double getMultiplier() {
 		return score.getMultiplier();
 	}
